Move code in 'toplevel' to api controller
